Move MAM view overrides and view-based code out of @converse/headless
updates #1352
